Big Update
Functions to upload the database is done.
