--- source/.env
@@ -1,7 +1,0 @@
-con = psycopg2.connect(
-    user = 'postgres',
-    password = '199811',
-    host = '127.0.0.1',
-    port = '5432',
-    database = 'pokedex'
-)
